Remove typetool.js and move two actions to tool/type.js
